feature: refactoring(change useMusic to useMusicPlayer)
